Cambio de tipo de dato de 'autor_id' a unsignedBigInteger en la migración de la tabla 'articulos'

+++ database/migrations/2025_01_06_130408_articulos.php
@@ -15,8 +15,10 @@ return new class extends Migration
             $table->id();
             $table->string("titulo");
             $table->string("contenido");
-            $table->integer("autor_id");
+            $table->unsignedBigInteger("autor_id");
             $table->timestamps();
+
+
         });
     }
